Move unspecified, but common subcommands to liboci-cli crate
The OCI Runtime Command Line Interface specification[0] specifies only five
subcommands to implement.  However, there are a number of other subcommands
that are commonly implemented by runtimes, and commonly used by their
clients.

Move the parsing of a number of those subcommands into the liboci-cli
crate.  This leaves only one subcommand "info" parsed in the youki crate.
As far as I can tell this is a youki specific extension, since it doesn't
appear in either runc or crun.

[0] https://github.com/opencontainers/runtime-tools/blob/master/docs/command-line-interface.md
