+++ crates/liboci-cli/src/events.rs
@@ -1,0 +1,15 @@
+use clap::Parser;
+
+/// Show resource statistics for the container
+#[derive(Parser, Debug)]
+pub struct Events {
+    /// Sets the stats collection interval in seconds (default: 5s)
+    #[clap(long, default_value = "5")]
+    pub interval: u32,
+    /// Display the container stats only once
+    #[clap(long)]
+    pub stats: bool,
+    /// Name of the container instance
+    #[clap(forbid_empty_values = true, required = true)]
+    pub container_id: String,
+}
