+++ crates/liboci-cli/src/info.rs
@@ -1,0 +1,5 @@
+use clap::Parser;
+
+/// Show information about the system
+#[derive(Parser, Debug)]
+pub struct Info {}
