+++ crates/liboci-cli/src/lib.rs
@@ -7,3 +7,18 @@ mod start;
 mod state;
 
 pub use {create::Create, delete::Delete, kill::Kill, start::Start, state::State};
+
+// Other common subcommands that aren't specified in the document
+mod events;
+mod exec;
+mod list;
+mod pause;
+mod ps;
+mod resume;
+mod run;
+mod spec;
+
+pub use {
+    events::Events, exec::Exec, list::List, pause::Pause, ps::Ps, resume::Resume, run::Run,
+    spec::Spec,
+};
