+++ crates/liboci-cli/src/list.rs
@@ -1,0 +1,5 @@
+use clap::Parser;
+
+/// List created containers
+#[derive(Parser, Debug)]
+pub struct List {}
