+++ crates/liboci-cli/src/pause.rs
@@ -1,0 +1,8 @@
+use clap::Parser;
+
+/// Suspend the processes within the container
+#[derive(Parser, Debug)]
+pub struct Pause {
+    #[clap(forbid_empty_values = true, required = true)]
+    pub container_id: String,
+}
