+++ crates/youki/src/commands/resume.rs
@@ -2,16 +2,10 @@
 use std::path::PathBuf;
 
 use anyhow::{Context, Result};
-use clap::Parser;
 
 use crate::commands::load_container;
 
-/// Resume the processes within the container
-#[derive(Parser, Debug)]
-pub struct Resume {
-    #[clap(forbid_empty_values = true, required = true)]
-    pub container_id: String,
-}
+use liboci_cli::Resume;
 
 // Resuming a container indicates resuming all processes in given container from paused state
 // This uses Freezer cgroup to suspend and resume processes
